requirements.txt
requests
beautifulsoup4
sqlalchemy
apscheduler
streamlit
python-dotenv
pandas
plotly
